@@ -1,59 +1,57 @@
 {
-  "name": "property-management-app",
-  "version": "1.0.0",
+  "name": "frontend-new",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
+    "dev": "vite",
+    "build": "tsc -b && vite build",
+    "lint": "eslint .",
+    "preview": "vite preview",
     "typecheck": "tsc --noEmit",
     "test": "vitest",
     "test:e2e": "playwright test",
     "test:e2e:ui": "playwright test --ui",
     "lhci": "lhci autorun",
-    "prepare": "husky"
+    "analyze:bundle": "ANALYZE=true npm run build",
+    "check:ci": "npm run lint && npm run typecheck && npm run test -- --run && npm run build"
   },
   "dependencies": {
-    "@headlessui/react": "^1.7.15",
-    "@hookform/resolvers": "^3.0.0",
-    "@tanstack/react-query": "^5.0.0",
-    "axios": "^1.12.0",
-    "lucide-react": "^0.276.0",
-    "next": "^14.2.33",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-hook-form": "^7.45.2",
-    "tailwindcss": "^3.3.0",
-    "zod": "^3.22.2"
+    "@tanstack/react-query": "^5.90.6",
+    "@types/leaflet": "^1.9.21",
+    "axios": "^1.13.2",
+    "leaflet": "^1.9.4",
+    "lucide-react": "^0.552.0",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-leaflet": "^5.0.0",
+    "react-router-dom": "^7.9.5"
   },
   "devDependencies": {
-    "@lhci/cli": "^0.13.0",
-    "@next/eslint-plugin-next": "^14.0.0",
-    "@playwright/test": "^1.48.0",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^14.0.0",
-    "@types/react": "^18.3.0",
-    "@types/react-dom": "^18.3.0",
-    "autoprefixer": "^10.4.14",
-    "axe-playwright": "^2.0.1",
-    "eslint": "^8.52.0",
-    "eslint-config-next": "^14.2.33",
-    "eslint-plugin-jsx-a11y": "^6.10.2",
-    "husky": "^9.1.7",
+    "@eslint/js": "^9.36.0",
+    "@lhci/cli": "^0.15.1",
+    "@playwright/test": "^1.56.1",
+    "@tailwindcss/postcss": "^4.1.16",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/node": "^24.6.0",
+    "@types/react": "^19.1.16",
+    "@types/react-dom": "^19.1.9",
+    "@vitejs/plugin-react": "^5.0.4",
+    "autoprefixer": "^10.4.21",
+    "axe-playwright": "^2.2.2",
+    "eslint": "^9.36.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.22",
+    "globals": "^16.4.0",
     "jsdom": "^27.1.0",
-    "lint-staged": "^16.2.6",
-    "msw": "^1.2.1",
-    "playwright": "^1.48.0",
-    "postcss": "^8.4.22",
-    "react-test-renderer": "^18.3.1",
-    "typescript": "^5.4.0",
-    "vitest": "^0.34.1"
-  },
-  "lint-staged": {
-    "*.{ts,tsx}": [
-      "eslint --fix",
-      "bash -c 'tsc --noEmit'"
-    ]
+    "postcss": "^8.5.6",
+    "rollup-plugin-visualizer": "^6.0.5",
+    "tailwindcss": "^4.1.16",
+    "typescript": "~5.9.3",
+    "typescript-eslint": "^8.45.0",
+    "vite": "^7.1.7",
+    "vitest": "^4.0.8"
   }
 }
