@@ -1,22 +1,63 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [
-    './app/**/*.{js,ts,jsx,tsx}',
-    './pages/**/*.{js,ts,jsx,tsx}',
-    './components/**/*.{js,ts,jsx,tsx}',
-    './_components/**/*.{js,ts,jsx,tsx}',
-    './_hooks/**/*.{js,ts,jsx,tsx}',
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {
       colors: {
-        primary: '#2563eb', // Tailwind blue-600
-        secondary: '#4f46e5', // Tailwind indigo-600
-        success: '#16a34a', // green-600
-        warning: '#d97706', // amber-600
-        danger: '#dc2626', // red-600
+        primary: {
+          DEFAULT: '#1E3A8A', // navy blue
+          50: '#EFF6FF',
+          100: '#DBEAFE',
+          200: '#BFDBFE',
+          300: '#93C5FD',
+          400: '#60A5FA',
+          500: '#3B82F6',
+          600: '#2563EB',
+          700: '#1E3A8A',
+          800: '#1E40AF',
+          900: '#1E3A8A',
+        },
+        accent: {
+          DEFAULT: '#0EA5E9', // teal blue
+          50: '#F0F9FF',
+          100: '#E0F2FE',
+          200: '#BAE6FD',
+          300: '#7DD3FC',
+          400: '#38BDF8',
+          500: '#0EA5E9',
+          600: '#0284C7',
+          700: '#0369A1',
+          800: '#075985',
+          900: '#0C4A6E',
+        },
+        brand: {
+          bg: '#FCF6EC',
+          panel: '#FFFFFF',
+          blue: '#BEE3F8',
+          teal: '#C6F6D5',
+          peach: '#FBD5D5',
+          text: '#0F172A',
+          subtle: '#64748B',
+          border: '#E5E7EB',
+        },
+      },
+      fontFamily: {
+        sans: ['Inter', 'system-ui', 'sans-serif'],
+      },
+      borderRadius: {
+        'card': '12px',
+        'input': '8px',
+        'xl': '12px',
+      },
+      boxShadow: {
+        'sm': '0 1px 2px rgba(0,0,0,0.05)',
+        'md': '0 2px 6px rgba(0,0,0,0.08)',
+        'card': '0 1px 2px rgba(16,24,40,.06), 0 1px 3px rgba(16,24,40,.08)',
       },
     },
   },
   plugins: [],
-};
+}
