@@ -1,5 +1,10 @@
+import { expect, afterEach } from 'vitest';
+import { cleanup } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import React from 'react';
 
-// Make React globally available for JSX transform
-globalThis.React = React;
+afterEach(() => {
+  cleanup();
+});
+
+// Make expect available globally
+global.expect = expect;
